Fixed recursion checking algorithm and improved debug statements
The new recursion algorithm handles test_recursive2, and backtracks on detecting a recursive
call. Adjacent subgoals to the relation are tried, and results from those relations are then
used when requerying the relation. This process is repeated until no new queries are added.
There is a minor (but non-trivial) performance hit. Instead of this algorithm, reordering the
GDL before recursion should be considered in the future.